Feat: 홈 화면 출, 퇴근 버튼 활성화
