+++ src/pages/MyPage/GoToWorkButton.jsx
@@ -1,24 +1,7 @@
-import axios from 'axios';
 import React from 'react';
-import { openModalWithMessage } from '../Modal/modalFunc';
+import { checkWorking } from './workingCheckFunc';
 
 const GoToWorkButton = () => {
-    const checkWorking = async() => {
-        console.log(window.localStorage.getItem("token"));
-        try{
-            const response = axios.post("http://localhost:8080/mypage/mainpage/attendance-go-work",{},{
-                headers: {
-                    'Authorization': window.localStorage.getItem("token")  // 여기에 토큰 추가
-                }
-            });
-            console.log(response);
-            openModalWithMessage("출근완료");
-        }
-        catch(err){
-            console.log(err);
-            openModalWithMessage("에러남");
-        }
-    }
     return (
         <button style={{
             width: '209px',
